fixed loops, need to "Redim" arrays
repaired the loops, now loop through without error and write to the arrays. 
arrays need to be rewritten to have no set length and instead "Redim" with preserve to add an additional spot to the end each time a new value needs to be written to insure the array is always the perfect length.
currently the loops work with array length of 10,000,000.